Remove StateRef from REF_CACHE on unmount
Prior to this commit, the `REF_CACHE` with
the default time-to-live of `0` retained the `StateRef`s
that were added to it even after the component
that called `useSwrv()` was unmounted (destroyed).
This allowed the `REF_CACHE` to grow indefinitely
as components that called `useSwrv()` were created
and destroyed.

This commit extends the callback passed to `onUnmounted()`
that removes the `StateRef` created by the current invocation
of `useSwrv()` from the array stored in the `REF_CACHE`
at the given key.