feat(*): update for vue 2.7 [KHCP-4402]
Update swrv for Vue 2.7 and updates the peerDependencies to Vue >= 2.7 < 3. Removes the peer
dependency for the @vue/composition-api plugin.

BREAKING CHANGE: SSR/Nuxt support is broken with this release. We are open to contributions to add
this feature back (the SSR code and tests remain in place and are commented out).